feat: move some types from @srcbook/web to @srcbook/components
