+++ packages/web/src/components/cells/generate-ai.tsx
@@ -4,7 +4,8 @@ import { useHotkeys } from 'react-hotkeys-hook';
 import { type CodeCellType, type MarkdownCellType } from '@srcbook/shared';
 import { generateCells } from '@/lib/server';
 import { CircleAlert, Trash2, Sparkles } from 'lucide-react';
-import { GenerateAICellType, SessionType } from '@/types';
+import { SessionType } from '@/types';
+import { GenerateAICellType } from '@srcbook/components/src/types';
 import { useCells } from '@srcbook/components/src/components/use-cell';
 import { cn } from '@/lib/utils';
 import { Button } from '@srcbook/components/src/components/ui/button';
